refactor into recording utils
