+++ pages/recording/[id].tsx
@@ -5,7 +5,8 @@ import { connect, ConnectedProps, useStore } from "react-redux";
 import { isTest } from "ui/utils/environment";
 import { getAccessibleRecording } from "ui/actions/session";
 import { Recording as RecordingInfo } from "ui/types";
-import { getRecordingMetadata, getRecordingURL, useGetRecordingId } from "ui/hooks/recordings";
+import { getRecordingMetadata, useGetRecordingId } from "ui/hooks/recordings";
+import { getRecordingURL } from "ui/utils/recording";
 import LoadingScreen from "ui/components/shared/LoadingScreen";
 import Upload from "./upload";
 import DevTools from "ui/components/DevTools";
