+++ src/ui/components/Events/ReplayInfo.tsx
@@ -2,14 +2,14 @@ import React, { ReactNode } from "react";
 import hooks from "ui/hooks";
 import { OperationsData } from "ui/types";
 import { formatRelativeTime } from "ui/utils/comments";
-import { getDisplayedUrl, getRecordingId } from "ui/utils/environment";
+import { getDisplayedUrl } from "ui/utils/environment";
 import { AvatarImage } from "../Avatar";
 import MaterialIcon from "../shared/MaterialIcon";
 import { getPrivacySummaryAndIcon } from "../shared/SharingModal/PrivacyDropdown";
 import { getUniqueDomains } from "../UploadScreen/Privacy";
 import { connect, ConnectedProps } from "react-redux";
 import * as actions from "ui/actions/app";
-import { showDurationWarning } from "ui/utils/recording";
+import { showDurationWarning, getRecordingId } from "ui/utils/recording";
 
 const Row = ({ children, onClick }: { children: ReactNode; onClick?: () => void }) => {
   const classes = "flex flex-row space-x-2 p-1.5 px-3 items-center text-left overflow-hidden";
