+++ src/ui/hooks/users.ts
@@ -1,9 +1,9 @@
 import { gql, useMutation, useQuery } from "@apollo/client";
 import { mutate, query } from "ui/utils/apolloClient";
 import { GET_USER_INFO, GET_USER_ID, DISMISS_NAG } from "ui/graphql/users";
+import { getRecordingId } from "ui/utils/recording";
 import { sendTelemetryEvent } from "ui/utils/telemetry";
 import { useGetRecording } from "./recordings";
-import { getRecordingId, isTest } from "ui/utils/environment";
 
 export async function getUserId() {
   const result = await query({
