+++ src/ui/utils/helpers.ts
@@ -1,3 +1,4 @@
+export const SLUG_SEPARATOR = "--";
 export function validateEmail(email: string) {
   const re =
     /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
@@ -41,7 +42,7 @@ export function extractIdAndSlug(params: string | string[] | undefined) {
     if (validateUUID(joined)) {
       id = joined;
     } else {
-      const parts = joined.split("--");
+      const parts = joined.split(SLUG_SEPARATOR);
       if (validateUUID(parts[1])) {
         slug = parts[0];
         id = parts[1];
